Moved SswConfig to config.rs
